requirements.txt
# LLM & AI
langchain==0.3.27
langchain-ollama==0.3.10
langchain-community==0.3.27
langgraph==0.2.54
google-generativeai==0.8.5
openai==2.8.0
ollama==0.6.1
torch==2.1.2
numba==0.58.1

# Audio & Speech
openai-whisper==20231117
sounddevice==0.5.3
soundfile==0.13.1
SpeechRecognition==3.14.3
pyttsx3==2.99
vosk==0.3.45

# Phone & SMS
twilio==9.8.8
flask==3.1.2

# Database
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11

# Jupyter & Notebooks
jupyterlab==4.4.10
ipykernel==7.1.0

# Utilities
deep-translator==1.11.4
langdetect==1.0.9
pandas==2.3.3
python-dotenv==1.2.1
requests==2.32.5
pydantic==2.12.4
pydantic-settings==2.12.0